Rename tooltip factory methods
